requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
numpy>=1.24.0

# V003: 云端大模型支持
openai>=1.0.0          # 通义千问兼容 OpenAI SDK
python-dotenv>=1.0.0   # 环境变量管理

# B: T7/T8 真实实现依赖
matplotlib>=3.8.0
seaborn>=0.13.0
python-docx>=1.1.0

# 测试脚本依赖（test_agent.py 使用）
httpx>=0.25.0

# RAG: 知识检索
chromadb>=0.5.0
pdfplumber>=0.10.0

# RAGAS: RAG 评估框架
ragas>=0.2.0
datasets>=2.14.0

